refactor: extract shared helper functions to iterm_mcpy/helpers.py
Prepares for splitting tool modules in subsequent tasks by moving
resolve_session and execute_* helpers to a shared location,
avoiding circular imports.
